+++ universe-client/src/pages/admin/SpeakersPage.jsx
@@ -1,0 +1,8 @@
+import React from "react";
+import SpeakersList from "@/components/admin/SpeakersList";
+
+const SpeakersPage = () => {
+  return <SpeakersList />;
+};
+
+export default SpeakersPage;